Linked users and Departments

+++ app/controllers/UsersController.php
@@ -99,9 +99,13 @@ class UsersController extends \BaseController {
 	 */
 	public function destroy($id)
 	{
+		if ($id == 1) {
+			return Response::json(['message' => 'error'], 400);
+		}
+
 		User::destroy($id);
 
-		return Redirect::route('admin.users.index');
+		return Response::json(['message' => 'deleted successfully'], 200);
 	}
 
 }
